% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{accuracy_pi_estimate}
\alias{accuracy_pi_estimate}
\title{Accuracy of pi estimate
The smaller the number the better.}
\usage{
accuracy_pi_estimate(pi_estimate)
}
\arguments{
\item{pi_estimate}{An estimation of pi.}
}
\value{
The absolute difference between the estimate and the true value of pi.
}
\description{
Accuracy of pi estimate
The smaller the number the better.
}
\examples{
accuracy_pi_estimate(3.14)
}
