% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{add_to_bottom}
\alias{add_to_bottom}
\title{Add rows to the bottom of a matrix
Used in rref.}
\usage{
add_to_bottom(A, to_append)
}
\arguments{
\item{A}{A numeric matrix.}

\item{to_append}{A vector of row indices to append to the bottom of the matrix.}
}
\value{
The matrix with the specified rows moved to the bottom.
}
\description{
Add rows to the bottom of a matrix
Used in rref.
}
\examples{
A <- matrix(1:9, nrow = 3)
add_to_bottom(A, 1)
}
