% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{adjugate}
\alias{adjugate}
\title{Create an Adjugate Matrix}
\usage{
adjugate(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
The adjugate of the matrix.
}
\description{
This function calculates the adjugate matrix of a square matrix. The adjugate matrix is the transpose of the cofactor matrix.
}
\examples{
A <- matrix(c(2, -1, -1, 2), nrow = 2)
adjugate(A)
}
