% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{approx_pi}
\alias{approx_pi}
\title{Approximate pi
Approximates pi using the formula 4 * points_within_circle / total_points.}
\usage{
approx_pi(dist)
}
\arguments{
\item{dist}{A logical vector indicating whether each point is within the unit circle.}
}
\value{
An approximation of pi.
}
\description{
Approximate pi
Approximates pi using the formula 4 * points_within_circle / total_points.
}
\examples{
pts <- generate_points(1000)
dist <- get_distance(pts)
approx_pi(dist)
}
