% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{approx_pi_resample}
\alias{approx_pi_resample}
\title{Approximate pi using resampling}
\usage{
approx_pi_resample(randVec)
}
\arguments{
\item{randVec}{A vector of random numbers.}
}
\value{
An approximation of pi.
}
\description{
Approximate pi using resampling
}
\examples{
ratios <- calc_ratio(100, 10, FALSE)
rand_beta <- generate_beta(ratios, 100, FALSE)
approx_pi_resample(rand_beta)
}
