% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{beta_mom}
\alias{beta_mom}
\title{Beta method of moments
from: https://stats.stackexchange.com/questions/376634/how-to-pick-starting-parameters-for-massfitdist-with-the-beta-distribution}
\usage{
beta_mom(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list with the shape1 and shape2 parameters of the beta distribution.
}
\description{
Beta method of moments
from: https://stats.stackexchange.com/questions/376634/how-to-pick-starting-parameters-for-massfitdist-with-the-beta-distribution
}
\examples{
ratios <- calc_ratio(100, 10, FALSE)
beta_mom(ratios)
}
