% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{calc_age}
\alias{calc_age}
\title{Calculate age}
\usage{
calc_age(birth_date, target_date, floor = TRUE)
}
\arguments{
\item{birth_date}{A character string representing the birth date in "YYYY-MM-DD" format.}

\item{target_date}{A character string representing the target date in "YYYY-MM-DD" format.}

\item{floor}{A logical indicating whether to floor the age to the nearest year.}
}
\value{
The age in years.
}
\description{
Calculate age
}
\examples{
calc_age(birth_date = "1989-01-24", target_date = "2009-01-21")
}
