% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{calc_ratio}
\alias{calc_ratio}
\title{Calculate ratio of points within}
\usage{
calc_ratio(n, samplingSize, plot)
}
\arguments{
\item{n}{The number of points to generate.}

\item{samplingSize}{The number of samples to take.}

\item{plot}{A logical indicating whether to plot a histogram of the ratios.}
}
\value{
A vector of ratios.
}
\description{
Calculate ratio of points within
}
\examples{
calc_ratio(100, 10, FALSE)
}
