% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{canonical_form}
\alias{canonical_form}
\title{Create canonical form (A = UDU^-1)}
\usage{
canonical_form(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
A list with the matrices U, D, and U_inv, representing the canonical form of A.
}
\description{
Create canonical form (A = UDU^-1)
}
\examples{
A <- matrix(c(2, -1, -1, 2), nrow = 2)
canonical_form(A)
}
