% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes.R
\name{coin_change_algo}
\alias{coin_change_algo}
\title{Coin change algorithm
TODO: This function is not yet implemented correctly.}
\usage{
coin_change_algo(coins, N)
}
\arguments{
\item{coins}{A vector of coin denominations.}

\item{N}{The target amount.}
}
\value{
The number of ways to make change for N.
}
\description{
Coin change algorithm
TODO: This function is not yet implemented correctly.
}
\examples{
coin_change_algo(c(1, 2), 4)
}
