% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{col_is_all_zero}
\alias{col_is_all_zero}
\title{Check if a column is all zero below the pivot
Used in rref.}
\usage{
col_is_all_zero(A, currentCol)
}
\arguments{
\item{A}{A numeric matrix.}

\item{currentCol}{The current column.}
}
\value{
A logical indicating whether the column is all zero below the pivot.
}
\description{
Check if a column is all zero below the pivot
Used in rref.
}
\examples{
A <- matrix(c(1, 0, 0, 1), nrow = 2)
col_is_all_zero(A, 1)
}
