% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{compare_floats}
\alias{compare_floats}
\title{Compare two floats}
\usage{
compare_floats(a, b, tol = 1e-06)
}
\arguments{
\item{a}{A float.}

\item{b}{A float.}

\item{tol}{The tolerance for the comparison.}
}
\value{
A logical indicating whether the floats are equal within the tolerance.
}
\description{
Compare two floats
}
\examples{
compare_floats(1.000001, 1)
}
