% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{concatenate_math}
\alias{concatenate_math}
\title{Concatenate numbers
Reverses the deconstruction of integers in digits.https://mathworld.wolfram.com/Concatenation.html}
\usage{
concatenate_math(x, y, base = 10)
}
\arguments{
\item{x}{The first number.}

\item{y}{The second number.}

\item{base}{The base to use for the concatenation.}
}
\value{
The concatenated number.
}
\description{
Concatenate numbers
Reverses the deconstruction of integers in digits.https://mathworld.wolfram.com/Concatenation.html
}
\examples{
concatenate_math(12, 34)
}
