% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_derivatives.R
\name{differentiate_polynomial}
\alias{differentiate_polynomial}
\title{Differentiate a polynomial}
\usage{
differentiate_polynomial(polynomial, order)
}
\arguments{
\item{polynomial}{A character string representing the polynomial.}

\item{order}{The order of the polynomial.}
}
\value{
A character string representing the derivative of the polynomial.
}
\description{
Differentiate a polynomial
}
\examples{
differentiate_polynomial("1+x+x^2+x^3+x^4", order = 4)
}
