% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes.R
\name{eratosthenes_sieve}
\alias{eratosthenes_sieve}
\title{Sieve of Erathostenes}
\usage{
eratosthenes_sieve(to)
}
\arguments{
\item{to}{upper bound of primes to return}
}
\value{
Vector of primes up to to-argument
}
\description{
Sieve of Erathostenes
}
\examples{
primes_to_11 <- eratosthenes_sieve(11)
}
