% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{extract_digits_matrix}
\alias{extract_digits_matrix}
\title{Extract digits from a vector of numbers and return a matrix}
\usage{
extract_digits_matrix(x)
}
\arguments{
\item{x}{A vector of numbers.}
}
\value{
A matrix of digits.
}
\description{
Extract digits from a vector of numbers and return a matrix
}
\examples{
extract_digits_matrix(c(123, 456))
}
