% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{extract_last_digits}
\alias{extract_last_digits}
\title{Extract the last digit of each number in a vector}
\usage{
extract_last_digits(x)
}
\arguments{
\item{x}{A vector of numbers.}
}
\value{
A vector of the last digits.
}
\description{
Extract the last digit of each number in a vector
}
\examples{
extract_last_digits(c(123, 456))
}
