% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{find_zero_vectors}
\alias{find_zero_vectors}
\title{Find zero vectors in a matrix}
\usage{
find_zero_vectors(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
A vector of indices of the zero vectors.
}
\description{
Find zero vectors in a matrix
}
\examples{
A <- matrix(c(1, 0, 0, 0), nrow = 2)
find_zero_vectors(A)
}
