% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{flip_sign}
\alias{flip_sign}
\title{Flip the sign of a scalar}
\usage{
flip_sign(scalar)
}
\arguments{
\item{scalar}{A scalar.}
}
\value{
The scalar with the sign flipped.
}
\description{
Flip the sign of a scalar
}
\examples{
flip_sign(-5)
}
