% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{gaussian_elimination}
\alias{gaussian_elimination}
\title{Perform Gaussian elimination
Used in rref.}
\usage{
gaussian_elimination(A, idxNonzero, currentCol)
}
\arguments{
\item{A}{A numeric matrix.}

\item{idxNonzero}{A vector of non-zero indices.}

\item{currentCol}{The current column.}
}
\value{
The matrix after Gaussian elimination.
}
\description{
Perform Gaussian elimination
Used in rref.
}
\examples{
A <- matrix(c(1, 2, 2, 4), nrow = 2, byrow = TRUE)
gaussian_elimination(A, 2, 1)
}
