% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{generalized_inverse}
\alias{generalized_inverse}
\title{Create a Generalized Inverse of a Matrix}
\usage{
generalized_inverse(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
The generalized inverse of the matrix.
}
\description{
This function calculates the generalized inverse (or Moore-Penrose inverse) of a matrix.
}
\examples{
A <- matrix(c(1, 2, 2, 4), nrow = 2, byrow = TRUE)
generalized_inverse(A)
}
