% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{generate_beta}
\alias{generate_beta}
\title{Generate beta distribution
Fits a beta distribution to a ratio data set and samples from this distribution.}
\usage{
generate_beta(ratioVector, outputLength, plot)
}
\arguments{
\item{ratioVector}{A vector of ratios.}

\item{outputLength}{The number of samples to generate.}

\item{plot}{A logical indicating whether to plot a histogram of the generated samples.}
}
\value{
A vector of samples from the beta distribution.
}
\description{
Generate beta distribution
Fits a beta distribution to a ratio data set and samples from this distribution.
}
\examples{
ratios <- calc_ratio(100, 10, FALSE)
generate_beta(ratios, 100, FALSE)
}
