% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{generate_gamma}
\alias{generate_gamma}
\title{Generate gamma distribution
Fits a gamma distribution to a ratio data set and samples from it.}
\usage{
generate_gamma(ratioVector, outputLength, plot)
}
\arguments{
\item{ratioVector}{A vector of ratios.}

\item{outputLength}{The number of samples to generate.}

\item{plot}{A logical indicating whether to plot a histogram of the generated samples.}
}
\value{
A vector of samples from the gamma distribution.
}
\description{
Generate gamma distribution
Fits a gamma distribution to a ratio data set and samples from it.
}
\examples{
ratios <- calc_ratio(100, 10, FALSE)
generate_gamma(ratios, 100, FALSE)
}
