% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{generate_points}
\alias{generate_points}
\title{Generate uniformly distributed points}
\usage{
generate_points(n)
}
\arguments{
\item{n}{The number of points to generate.}
}
\value{
A matrix of uniformly distributed points.
}
\description{
Generate uniformly distributed points
}
\examples{
generate_points(10)
}
