% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{get_beta_dist}
\alias{get_beta_dist}
\title{Get beta distribution parameters
Fits a beta distribution to a ratio data set and returns the parameters.}
\usage{
get_beta_dist(ratioVector)
}
\arguments{
\item{ratioVector}{A vector of ratios.}
}
\value{
The parameters of the beta distribution.
}
\description{
Get beta distribution parameters
Fits a beta distribution to a ratio data set and returns the parameters.
}
\examples{
ratios <- calc_ratio(100, 10, FALSE)
get_beta_dist(ratios)
}
