% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{get_digits}
\alias{get_digits}
\title{Get number of digits of a number
log10 of a number gives us the approximate number of digits - 1. By using the floor operation we cancel the decimal noise}
\usage{
get_digits(x, base = 10)
}
\arguments{
\item{x}{A positive integer.}

\item{base}{The base to use for the logarithm.}
}
\value{
The number of digits in x.
}
\description{
Get number of digits of a number
log10 of a number gives us the approximate number of digits - 1. By using the floor operation we cancel the decimal noise
}
\examples{
get_digits(123)
}
