% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{get_distance}
\alias{get_distance}
\title{Get distance
Calculates the vector norm and checks whether the point is within the radius (i.e. < 1).}
\usage{
get_distance(XY)
}
\arguments{
\item{XY}{A matrix of points.}
}
\value{
A logical vector indicating whether each point is within the unit circle.
}
\description{
Get distance
Calculates the vector norm and checks whether the point is within the radius (i.e. < 1).
}
\examples{
pts <- generate_points(10)
get_distance(pts)
}
