% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{get_powers_of_ten}
\alias{get_powers_of_ten}
\title{Get powers of ten}
\usage{
get_powers_of_ten(ndigits, include_one = FALSE, full_seq = TRUE)
}
\arguments{
\item{ndigits}{The number of digits.}

\item{include_one}{A logical indicating whether to include 1 in the sequence.}

\item{full_seq}{A logical indicating whether to return the full sequence or just the upper bound.}
}
\value{
A vector of powers of ten.
}
\description{
Get powers of ten
}
\examples{
get_powers_of_ten(3)
}
