% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{inverse}
\alias{inverse}
\title{Inverse of a square matrix}
\usage{
inverse(A)
}
\arguments{
\item{A}{A square numeric matrix.}
}
\value{
The inverse of the matrix.
}
\description{
Inverse of a square matrix
}
\examples{
A <- matrix(c(2, -1, -1, 2), nrow = 2)
inverse(A)
}
