% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{is_defective}
\alias{is_defective}
\title{Check if a Matrix is Defective}
\usage{
is_defective(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
A logical value indicating whether the matrix is defective (\code{TRUE}) or not (\code{FALSE}).
}
\description{
This function checks if a given square matrix is defective. A matrix is defective if the geometric multiplicity of any of its eigenvalues is less than the algebraic multiplicity.
}
\examples{
# A defective matrix
A <- matrix(c(1, 1, 0, 1), nrow = 2)
is_defective(A)

# A non-defective matrix
B <- matrix(c(1, 0, 0, 2), nrow = 2)
is_defective(B)
}
