% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{is_divisible_by_digit}
\alias{is_divisible_by_digit}
\title{Check if a number is divisible by its digits
https://www.youtube.com/watch?v=UDQjn_-pDSs}
\usage{
is_divisible_by_digit(x)
}
\arguments{
\item{x}{A number.}
}
\value{
A logical indicating whether the number is divisible by its digits.
}
\description{
Check if a number is divisible by its digits
https://www.youtube.com/watch?v=UDQjn_-pDSs
}
\examples{
is_divisible_by_digit(12)
}
