% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{is_pos_def}
\alias{is_pos_def}
\title{Check if a matrix is positive definite}
\usage{
is_pos_def(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
A logical indicating whether the matrix is positive definite.
}
\description{
Check if a matrix is positive definite
}
\examples{
A <- matrix(c(2, -1, -1, 2), nrow = 2)
is_pos_def(A)
}
