% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{lin_dep_Cautchy_Schwartz}
\alias{lin_dep_Cautchy_Schwartz}
\title{Cauchy-Schwarz inequality test for linear dependence
Single Cauchy-Schwarz inequality test. Check two vectors.}
\usage{
lin_dep_Cautchy_Schwartz(u, v, tol = 1e-05)
}
\arguments{
\item{u}{A numeric vector.}

\item{v}{A numeric vector.}

\item{tol}{The tolerance for the comparison.}
}
\value{
A logical indicating whether the vectors are linearly dependent.
}
\description{
Cauchy-Schwarz inequality test for linear dependence
Single Cauchy-Schwarz inequality test. Check two vectors.
}
\examples{
u <- c(1, 2)
v <- c(2, 4)
lin_dep_Cautchy_Schwartz(u, v)
}
