% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{lin_dep_Cautchy_Schwartz_matrix}
\alias{lin_dep_Cautchy_Schwartz_matrix}
\title{Cauchy-Schwarz inequality test for linear dependence in a matrix
Cauchy-Schwarz-Inequality to identify parallel cols/rows. Check for more than
two vecs}
\usage{
lin_dep_Cautchy_Schwartz_matrix(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
A data frame with the indices of the linearly dependent rows/cols, or NULL if none are found.
}
\description{
Cauchy-Schwarz inequality test for linear dependence in a matrix
Cauchy-Schwarz-Inequality to identify parallel cols/rows. Check for more than
two vecs
}
\examples{
A <- matrix(c(1, 2, 2, 4), nrow = 2, byrow = TRUE)
lin_dep_Cautchy_Schwartz_matrix(A)
}
