% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{mat_pow}
\alias{mat_pow}
\title{Calculate matrix powers}
\usage{
mat_pow(A, n)
}
\arguments{
\item{A}{A numeric matrix.}

\item{n}{The power to raise the matrix to.}
}
\value{
The matrix A raised to the power of n.
}
\description{
Calculate matrix powers
}
\examples{
A <- matrix(c(2, -1, -1, 2), nrow = 2)
mat_pow(A, 2)
}
