% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_derivatives.R
\name{matrix_derivative}
\alias{matrix_derivative}
\title{Title}
\usage{
matrix_derivative(polyVec, d_dx, order)
}
\arguments{
\item{order}{}
}
\description{
Title
}
\examples{
differentiate_polynomial("1+x+x^2+x^3+x^4", order = 4)
differentiate_polynomial("1+x^9+3x^3+5x^6", order = 9)
differentiate_polynomial("0+x+x^2", order = 2)
}
