% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{MCMC_h_Pi}
\alias{MCMC_h_Pi}
\title{Metropolis-Hastings-like algorithm for pi estimation
Same as MCMC_Pi but with a real Hastings ratio.}
\usage{
MCMC_h_Pi(nInit = 1e+06, samplingSize = 10000, nSD = 1000, nIter)
}
\arguments{
\item{nInit}{The number of initial points to generate for the prior distribution.}

\item{samplingSize}{The number of samples to take for the prior distribution.}

\item{nSD}{The number of points to generate for the standard deviation calculation.}

\item{nIter}{The number of iterations for the MCMC chain.}
}
\value{
A vector representing the MCMC chain.
}
\description{
Metropolis-Hastings-like algorithm for pi estimation
Same as MCMC_Pi but with a real Hastings ratio.
}
\examples{
MCMC_h_Pi(nIter = 100)
}
