% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{MCMC_Pi}
\alias{MCMC_Pi}
\title{MCMC-like algorithm for pi estimation
MCMC-like algorithm, but instead of using a hastings ratio, I used the accuracy measure,
which kinda breaks the purpose of the MCMC, which is used when the true value is unknown.
But heck, its just for fun, right.}
\usage{
MCMC_Pi(nInit = 1e+06, samplingSize = 10000, nSD = 1000, nIter)
}
\arguments{
\item{nInit}{The number of initial points to generate for the prior distribution.}

\item{samplingSize}{The number of samples to take for the prior distribution.}

\item{nSD}{The number of points to generate for the standard deviation calculation.}

\item{nIter}{The number of iterations for the MCMC chain.}
}
\value{
A vector representing the MCMC chain.
}
\description{
MCMC-like algorithm for pi estimation
MCMC-like algorithm, but instead of using a hastings ratio, I used the accuracy measure,
which kinda breaks the purpose of the MCMC, which is used when the true value is unknown.
But heck, its just for fun, right.
}
\examples{
MCMC_Pi(nIter = 100)
}
