% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{mean_estimate}
\alias{mean_estimate}
\title{Mean estimate of pi
Descriptive statistics of the pi estimate collected from several iterations.}
\usage{
mean_estimate(
  n,
  nIter,
  type,
  samplingSize = 10000,
  outputLength = 1e+06,
  distr = "beta"
)
}
\arguments{
\item{n}{The number of points to generate.}

\item{nIter}{The number of iterations.}

\item{type}{The type of estimation ("empirical" or "resampled").}

\item{samplingSize}{The number of samples to take for the ratio calculation.}

\item{outputLength}{The number of samples to generate from the distribution.}

\item{distr}{The distribution to use for resampling ("beta" or "gamma").}
}
\value{
A data frame with descriptive statistics of the pi estimates.
}
\description{
Mean estimate of pi
Descriptive statistics of the pi estimate collected from several iterations.
}
\examples{
mean_estimate(100, 10, "empirical")
}
