% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{orthogonalize}
\alias{orthogonalize}
\title{Orthogonalize a matrix
Orthogonalize matrix (AAT or ATA)}
\usage{
orthogonalize(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
An orthogonalized matrix.
}
\description{
Orthogonalize a matrix
Orthogonalize matrix (AAT or ATA)
}
\examples{
A <- matrix(c(1, 2, 2, 4), nrow = 2, byrow = TRUE)
orthogonalize(A)
}
