% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{plot_all_digits}
\alias{plot_all_digits}
\title{Plot the distribution of all digits in a vector of numbers}
\usage{
plot_all_digits(vec)
}
\arguments{
\item{vec}{A numeric vector.}
}
\value{
A ggplot object representing the distribution of the digits.
}
\description{
This function takes a vector of numbers, extracts all the digits,
and plots their distribution as a bar chart.
}
\examples{
plot_all_digits(c(123, 456, 789, 258))
}
