% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{plot_eigenvec}
\alias{plot_eigenvec}
\title{Plot eigenvectors of a 2x2 matrix}
\usage{
plot_eigenvec(
  A,
  offset = 1,
  plotBasisVecs = T,
  plotSpan = T,
  plotTransBasis = T
)
}
\arguments{
\item{A}{A 2x2 numeric matrix.}

\item{offset}{The offset for the plot limits.}

\item{plotBasisVecs}{A logical indicating whether to plot the basis vectors.}

\item{plotSpan}{A logical indicating whether to plot the span of the vectors.}

\item{plotTransBasis}{A logical indicating whether to plot the transformed basis vectors.}
}
\value{
A plot of the eigenvectors.
}
\description{
Plot eigenvectors of a 2x2 matrix
}
\examples{
A <- matrix(c(2, -1, -1, 2), nrow = 2)
plot_eigenvec(A)
}
