% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{plot_matrix_transformation}
\alias{plot_matrix_transformation}
\title{Plot Matrix Transformation}
\usage{
plot_matrix_transformation(
  A,
  v,
  offset = 1,
  plotBasisVecs = TRUE,
  splitPlot = TRUE
)
}
\arguments{
\item{A}{A 2x2 numeric matrix.}

\item{v}{A numeric vector.}

\item{offset}{A numeric value specifying the offset for the plot limits. Defaults to \code{1}.}

\item{plotBasisVecs}{A logical value indicating whether to plot the basis vectors. Defaults to \code{TRUE}.}

\item{splitPlot}{A logical value indicating whether to create a split plot showing the transformation before and after. Defaults to \code{TRUE}.}
}
\value{
A plot of the matrix transformation.
}
\description{
This function visualizes the transformation of a vector \code{v} by a 2x2 matrix \code{A}.
}
\examples{
A <- matrix(c(2, -1, -1, 2), nrow = 2)
v <- c(1, 1)
plot_matrix_transformation(A, v)
}
