% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{ref}
\alias{ref}
\title{Get row echelon form of a matrix}
\usage{
ref(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
The row echelon form of the matrix.
}
\description{
Get row echelon form of a matrix
}
\examples{
A <- matrix(c(1, 2, 2, 4), nrow = 2, byrow = TRUE)
ref(A)
}
