% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{remove_parallel_vectors}
\alias{remove_parallel_vectors}
\title{Remove parallel vectors from a matrix}
\usage{
remove_parallel_vectors(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
The matrix with parallel vectors removed.
}
\description{
Remove parallel vectors from a matrix
}
\examples{
A <- matrix(c(1, 2, 2, 4), nrow = 2, byrow = TRUE)
remove_parallel_vectors(A)
}
