% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{scoring}
\alias{scoring}
\title{Scoring function
The lower the absolute difference between pi and the estimate the better.}
\usage{
scoring(res)
}
\arguments{
\item{res}{The result of the accuracy_pi_estimate function.}
}
\value{
A score from 0 to 6.
}
\description{
Scoring function
The lower the absolute difference between pi and the estimate the better.
}
\examples{
scoring(0.05)
}
