% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes.R
\name{sieve_ov_Erathostenes}
\alias{sieve_ov_Erathostenes}
\title{Sieve of Erathostenes}
\usage{
sieve_ov_Erathostenes(to)
}
\arguments{
\item{to}{upper bound of primes to return}
}
\value{
Vector of primes up to to-argument
}
\description{
Sieve of Erathostenes
}
\examples{
primes_to_11 <- sieve_ov_Erathostenes(11)
}
