% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{swap}
\alias{swap}
\title{Swap rows or columns of a matrix}
\usage{
swap(A, old, new, col = TRUE)
}
\arguments{
\item{A}{A numeric matrix.}

\item{old}{The index of the row/column to swap.}

\item{new}{The index of the row/column to swap with.}

\item{col}{A logical indicating whether to swap columns instead of rows.}
}
\value{
The matrix with the rows/columns swapped.
}
\description{
Swap rows or columns of a matrix
}
\examples{
A <- matrix(1:4, nrow = 2)
swap(A, 1, 2)
}
