% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{swap_zero_vectors}
\alias{swap_zero_vectors}
\title{Swap zero vectors to the bottom of a matrix
Used to put zero vectors at the bottom if found in matrix.}
\usage{
swap_zero_vectors(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
The matrix with the zero vectors at the bottom.
}
\description{
Swap zero vectors to the bottom of a matrix
Used to put zero vectors at the bottom if found in matrix.
}
\examples{
A <- matrix(c(1, 0, 0, 0), nrow = 2, byrow = TRUE)
swap_zero_vectors(A)
}
