% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{test_accuracy}
\alias{test_accuracy}
\title{Test accuracy of pi estimation
Compare time and score of accuracy. Not for MCMC as we only consider the last values.}
\usage{
test_accuracy(
  n,
  nIter,
  type,
  samplingSize = 10000,
  outputLength = 1e+06,
  distr = "beta"
)
}
\arguments{
\item{n}{The number of points to generate.}

\item{nIter}{The number of iterations.}

\item{type}{The type of estimation ("empirical" or "resampled").}

\item{samplingSize}{The number of samples to take for the ratio calculation.}

\item{outputLength}{The number of samples to generate from the distribution.}

\item{distr}{The distribution to use for resampling ("beta" or "gamma").}
}
\value{
A data frame with the accuracy scores.
}
\description{
Test accuracy of pi estimation
Compare time and score of accuracy. Not for MCMC as we only consider the last values.
}
\examples{
test_accuracy(100, 10, "empirical")
}
