% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{transform_to_base_ten}
\alias{transform_to_base_ten}
\title{Transform to base ten}
\usage{
transform_to_base_ten(x, base)
}
\arguments{
\item{x}{The number to transform.}

\item{base}{The base of the number.}
}
\value{
The number in base ten.
}
\description{
Transform to base ten
}
\examples{
transform_to_base_ten(1000, 2)
}
