% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{adjugate}
\alias{adjugate}
\title{Create an adjugate matrix
Create an adjugate matrix (transpose of a cofactor matrix)}
\usage{
adjugate(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
The adjugate of the matrix.
}
\description{
Create an adjugate matrix
Create an adjugate matrix (transpose of a cofactor matrix)
}
\examples{
A <- matrix(c(2, -1, -1, 2), nrow = 2)
adjugate(A)
}
