% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{canonical_form}
\alias{canonical_form}
\title{Create Canonical Form (Diagonalization) of a Matrix}
\usage{
canonical_form(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
A list containing the matrices \code{U}, \code{D}, and \code{U_inv}.
}
\description{
This function computes the canonical form (or diagonalization) of a square matrix \code{A}. The canonical form is represented by the equation \code{A = UDU^-1}, where \code{U} is the matrix of eigenvectors, \code{D} is the diagonal matrix of eigenvalues, and \code{U_inv} is the inverse of \code{U}.
}
\examples{
A <- matrix(c(2, -1, -1, 2), nrow = 2)
canonical_form(A)
}
