% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{check_Penrose_cond}
\alias{check_Penrose_cond}
\title{Check Penrose conditions of a generalized inverse}
\usage{
check_Penrose_cond(A, G, all_Penrose_check = F, digits = 2)
}
\arguments{
\item{A}{A numeric matrix.}

\item{G}{The generalized inverse of A.}

\item{all_Penrose_check}{A logical indicating whether to check all four Penrose conditions.}

\item{digits}{The number of digits to round to.}
}
\value{
A logical indicating whether the Penrose conditions are met.
}
\description{
Check Penrose conditions of a generalized inverse
}
\examples{
A <- matrix(c(1, 2, 2, 4), nrow = 2, byrow = TRUE)
G <- generalized_inverse(A)
check_Penrose_cond(A, G)
}
