% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{create_basis}
\alias{create_basis}
\title{Create a basis
Create linear independent vectors aka a basis}
\usage{
create_basis(dim, negative = TRUE, upper = 9, returnMat = FALSE)
}
\arguments{
\item{dim}{The dimension of the basis.}

\item{negative}{A logical indicating whether to include negative values.}

\item{upper}{The upper bound for the random numbers.}

\item{returnMat}{A logical indicating whether to return a matrix or a list of vectors.}
}
\value{
A basis, either as a matrix or a list of vectors.
}
\description{
Create a basis
Create linear independent vectors aka a basis
}
\examples{
create_basis(3)
}
