% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{estimate_pi_empirical}
\alias{estimate_pi_empirical}
\title{Estimate pi empirically
The empirical version just generates data between 0 and 1 and calculates the
distance to the radius of the unit circle.
More details here : https://en.wikipedia.org/wiki/Approximations_of_\%CF\%80#Summing_a_circle's_area}
\usage{
estimate_pi_empirical(n)
}
\arguments{
\item{n}{The number of points to generate.}
}
\value{
An approximation of pi.
}
\description{
Estimate pi empirically
The empirical version just generates data between 0 and 1 and calculates the
distance to the radius of the unit circle.
More details here : https://en.wikipedia.org/wiki/Approximations_of_\%CF\%80#Summing_a_circle's_area
}
\examples{
estimate_pi_empirical(1000)
}
