% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pi.R
\name{estimate_pi_resampled}
\alias{estimate_pi_resampled}
\title{Estimate pi using resampling
This function uses a probabilistic approach. The ratio of points that are
within the radius is sampled from a fitted gamma distribution. Currently, the
mean of randomly generated data from this vector is used to estimate pi.}
\usage{
estimate_pi_resampled(
  n,
  outputLength = 1e+06,
  samplingSize = 1e+05,
  distr = "beta",
  plot = FALSE
)
}
\arguments{
\item{n}{The number of points to generate.}

\item{outputLength}{The number of samples to generate from the distribution.}

\item{samplingSize}{The number of samples to take for the ratio calculation.}

\item{distr}{The distribution to use for resampling ("beta" or "gamma").}

\item{plot}{A logical indicating whether to plot a histogram of the generated samples.}
}
\value{
An approximation of pi.
}
\description{
Estimate pi using resampling
This function uses a probabilistic approach. The ratio of points that are
within the radius is sampled from a fitted gamma distribution. Currently, the
mean of randomly generated data from this vector is used to estimate pi.
}
\examples{
estimate_pi_resampled(100, 100, 10)
}
