% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{extract_digits}
\alias{extract_digits}
\title{Extract digits from a vector of numbers}
\usage{
extract_digits(x_vec)
}
\arguments{
\item{x_vec}{A vector of numbers.}
}
\value{
A vector of digits.
}
\description{
Extract digits from a vector of numbers
}
\examples{
extract_digits(c(123, 456))
}
