% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{fast_exp}
\alias{fast_exp}
\title{Fast exponentiation using canonical form}
\usage{
fast_exp(A, p)
}
\arguments{
\item{A}{A numeric matrix.}

\item{p}{The power to raise the matrix to.}
}
\value{
The matrix A raised to the power of p.
}
\description{
Fast exponentiation using canonical form
}
\examples{
A <- matrix(c(2, -1, -1, 2), nrow = 2)
fast_exp(A, 2)
}
