% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes.R
\name{get_circular_primes}
\alias{get_circular_primes}
\title{Get circular primes
E.g. 113, 131, and 311, i.e. circular permutatios that are also primes}
\usage{
get_circular_primes(primes)
}
\arguments{
\item{primes}{A vector of primes.}
}
\value{
A vector of circular primes.
}
\description{
Get circular primes
E.g. 113, 131, and 311, i.e. circular permutatios that are also primes
}
\examples{
primes_to_100 <- eratosthenes_sieve(100)
get_circular_primes(primes_to_100)
}
