% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{lin_dep_Cautchy_Schwartz}
\alias{lin_dep_Cautchy_Schwartz}
\title{Cauchy-Schwarz Inequality Test for Linear Dependence}
\usage{
lin_dep_Cautchy_Schwartz(u, v, tol = 1e-05)
}
\arguments{
\item{u}{A numeric vector.}

\item{v}{A numeric vector.}

\item{tol}{A numeric value specifying the tolerance for the comparison. Defaults to \code{1e-05}.}
}
\value{
A logical value indicating whether the vectors are linearly dependent (\code{TRUE}) or not (\code{FALSE}).
}
\description{
This function uses the Cauchy-Schwarz inequality to test if two vectors are linearly dependent.
}
\examples{
u <- c(1, 2)
v <- c(2, 4)
lin_dep_Cautchy_Schwartz(u, v)
}
