% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{lin_dep_Cautchy_Schwartz_matrix}
\alias{lin_dep_Cautchy_Schwartz_matrix}
\title{Cauchy-Schwarz Inequality Test for Linear Dependence in a Matrix}
\usage{
lin_dep_Cautchy_Schwartz_matrix(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
A data frame with the indices of the linearly dependent rows or columns, or \code{NULL} if none are found.
}
\description{
This function uses the Cauchy-Schwarz inequality to identify linearly dependent rows or columns in a matrix.
}
\examples{
A <- matrix(c(1, 2, 2, 4), nrow = 2, byrow = TRUE)
lin_dep_Cautchy_Schwartz_matrix(A)
}
