% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_list_comprehension.R
\name{list_comp_2d}
\alias{list_comp_2d}
\title{2D List Comprehension
Emulates 2D list comprehension from Python.}
\usage{
list_comp_2d(row, col, x = NULL, cond = NULL)
}
\arguments{
\item{row}{Number of rows}

\item{col}{Number of columns}

\item{x}{Expression to evaluate for each element}

\item{cond}{Condition to filter elements}
}
\value{
A matrix
}
\description{
2D List Comprehension
Emulates 2D list comprehension from Python.
}
\examples{
# Default value is from 1 to row*col
list_comp_2d(row = 3, col = 3)
# zero matrix
list_comp_2d(row = 3, col = 3, x = 0)
# tetration
list_comp_2d(row = 3, col = 3, x = "i^i")
# ln(x)
list_comp_2d(row = 3, col = 3, x = "log(i)")
# cond test 1
list_comp_2d(row = 3, col = 3, cond = "i \%\% 2 == 0")
# cond test 2
list_comp_2d(row = 3, col = 3, cond = "i != 3 & i != 5")
}
