% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_derivatives.R
\name{matrix_derivative}
\alias{matrix_derivative}
\title{Differentiate a polynomial using a derivative matrix}
\usage{
matrix_derivative(polyVec, d_dx, order)
}
\arguments{
\item{polyVec}{A numeric vector representing the polynomial.}

\item{d_dx}{The derivative matrix.}

\item{order}{The order of the polynomial.}
}
\value{
A character string representing the derivative of the polynomial.
}
\description{
Differentiate a polynomial using a derivative matrix
}
\examples{
d_dx <- construct_derivate_matrix(4)
polyVec <- prep_polynomial("1+x+x^2+x^3+x^4", 4)
matrix_derivative(polyVec, d_dx, 4)
}
