% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{orthogonalize}
\alias{orthogonalize}
\title{Orthogonalize a Matrix}
\usage{
orthogonalize(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
An orthogonalized matrix.
}
\description{
This function orthogonalizes a matrix by computing the eigenvectors of \code{A \%*\% t(A)} or \code{t(A) \%*\% A}.
}
\examples{
A <- matrix(c(1, 2, 2, 4), nrow = 2, byrow = TRUE)
orthogonalize(A)
}
