% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{plot_eigenvec}
\alias{plot_eigenvec}
\title{Plot Eigenvectors of a 2x2 Matrix}
\usage{
plot_eigenvec(
  A,
  offset = 1,
  plotBasisVecs = TRUE,
  plotSpan = TRUE,
  plotTransBasis = TRUE
)
}
\arguments{
\item{A}{A 2x2 numeric matrix.}

\item{offset}{A numeric value specifying the offset for the plot limits. Defaults to \code{1}.}

\item{plotBasisVecs}{A logical value indicating whether to plot the basis vectors. Defaults to \code{TRUE}.}

\item{plotSpan}{A logical value indicating whether to plot the span of the vectors. Defaults to \code{TRUE}.}

\item{plotTransBasis}{A logical value indicating whether to plot the transformed basis vectors. Defaults to \code{TRUE}.}
}
\value{
A plot of the eigenvectors.
}
\description{
This function visualizes the eigenvectors of a 2x2 matrix, showing the transformation of the basis vectors and the span of the eigenvectors.
}
\examples{
A <- matrix(c(2, -1, -1, 2), nrow = 2)
plot_eigenvec(A)
}
