% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{plot_matrix_transformation}
\alias{plot_matrix_transformation}
\title{Plot matrix transformation Ax = y}
\usage{
plot_matrix_transformation(
  A,
  v,
  offset = 1,
  plotBasisVecs = TRUE,
  splitPlot = TRUE
)
}
\arguments{
\item{A}{A 2x2 numeric matrix.}

\item{v}{A numeric vector.}

\item{offset}{The offset for the plot limits.}

\item{plotBasisVecs}{A logical indicating whether to plot the basis vectors.}

\item{splitPlot}{A logical indicating whether to split the plot into two.}
}
\value{
A plot of the matrix transformation.
}
\description{
Plot matrix transformation Ax = y
}
\examples{
A <- matrix(c(2, -1, -1, 2), nrow = 2)
v <- c(1, 1)
plot_matrix_transformation(A, v)
}
