% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{rank_matrix}
\alias{rank_matrix}
\title{Find the rank of a matrix}
\usage{
rank_matrix(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
The rank of the matrix.
}
\description{
Find the rank of a matrix
}
\examples{
A <- matrix(c(1, 2, 2, 4), nrow = 2, byrow = TRUE)
rank_matrix(A)
}
