% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{rank_square_matrix}
\alias{rank_square_matrix}
\title{Find the rank of a square matrix}
\usage{
rank_square_matrix(A, tol = 1e-12)
}
\arguments{
\item{A}{A square numeric matrix.}

\item{tol}{The tolerance for the eigenvalues.}
}
\value{
The rank of the matrix.
}
\description{
Find the rank of a square matrix
}
\examples{
A <- matrix(c(1, 2, 2, 4), nrow = 2, byrow = TRUE)
rank_square_matrix(A)
}
