% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{rref}
\alias{rref}
\title{Get Reduced Row Echelon Form of a Matrix}
\usage{
rref(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
The reduced row echelon form of the matrix.
}
\description{
This function converts a matrix to its reduced row echelon form using Gaussian elimination.
}
\examples{
A <- matrix(c(1, 2, 2, 4), nrow = 2, byrow = TRUE)
rref(A)
}
