% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafnaf.R
\name{singular_value_decomposition}
\alias{singular_value_decomposition}
\title{Singular value decomposition}
\usage{
singular_value_decomposition(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
A list with the matrices P, D, and Q, representing the singular value decomposition of A.
}
\description{
Singular value decomposition
}
\examples{
A <- matrix(c(1, 2, 2, 4), nrow = 2, byrow = TRUE)
singular_value_decomposition(A)
}
